feat(tasks): create searchUsersMenu shared component
